Cleanup code and remove duplicate implementations
