+++ libs/api/src/main/java/de/cyface/api/MeasurementRetriever.java
@@ -1,9 +1,20 @@
 /*
- * Copyright (C) 2020 Cyface GmbH - All Rights Reserved
+ * Copyright 2020-2021 Cyface GmbH
  *
- * This file is part of the Cyface Server Backend.
- * Unauthorized copying of this file, via any medium is strictly prohibited
- * Proprietary and confidential
+ * This file is part of the Cyface Data Collector.
+ *
+ * The Cyface Data Collector is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The Cyface Data Collector is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with the Cyface Data Collector. If not, see <http://www.gnu.org/licenses/>.
  */
 package de.cyface.api;
